'security' is the best for checking existence
`git-credential` is best for writing, but it isn't good at checking for
the existence of a credential in the keychain. The only command that
could be used for this is 'fill', but if you attempt to fill a password
that isn't in the keychain, the user is shown a graphical prompt which
asks them to supply the password -- it's unclear when this would ever be
the right use case for khef.